Format a failed test message correctly when the test uses #[test] rather than #[google_test].
The Rust test harness uses std::fmt::Debug to output the contents of the Err variant of a Result when the test fails by returning Result::Err. This does not pretty-print the failure message as intended.

This change adds a custom Debug implementation for TestAssertionFailure which defers to the Display implementation. The failure message is then correct.

PiperOrigin-RevId: 494152135