Support an arbitrary matcher as an argument to contains(...).times(...).
This allows expressions which match on broader conditions on the number of times matching items appear in a container. Previously, one had to specify the exact number of items. Now one can, for example, state only that at least or at most some number of matching items must be present.

PiperOrigin-RevId: 497143881