Introduce assert_that and assert_pred macros which panic when the matcher does not match, respectively the predicate is false.
This is analogous to other Rust test libraries, so may be more comfortable for some developers.

Also fixes a bug in the implementation of verify_pred in which the macro was not being referenced from the source crate. This would break assert_pred in any code not in the googletest crate.

PiperOrigin-RevId: 497144231