Move the integration tests from googletest/tests to googletest/integration_tests.
This disables their automatic discovery by Cargo. Automatic discovery by Cargo is problematic because some of the dependent test binaries of these integration tests must be compiled in the test harness but then treated as ordinary binaries so that the integration tests find them. This is not possible in Cargo currently, so recent changes have caused the integration tests in Cargo to fail.

This change effectively disables the integration tests in Cargo for the time being. A follow-up change will add a shell script which will use cargo rustc -- --test to compile the integration tests and their dependent binaries and then run them. This simulates cargo test and should suffice for CI purposes.

PiperOrigin-RevId: 497165316