Modify the test helper binary simple_assertion_failure to be compiled with the standard Rust test harness.
Thus the binary is compiled with --test and no longer needs a main() function which invokes the test method. The test method is now annotatd with #[test] rather than #[google_test_wrapper] so that it runs as an ordinary test. The test itself in lib.rs asserts on the console output as before.

The test method is marked as ignored so that it is not automatically run by Cargo when testing the full library. The test itself overrides this by running the test binary with the --ignored flag.

In follow-up CLs, all test helper binaries will be adjusted to run in the test harness. For this reason, this change already invokes all binaries with the --ignored flag. The macro google_test_wrapper can then be folded into google_test, simplfying the exposed API surface of the googletest_macro crate.

PiperOrigin-RevId: 497165449