Fix bulding the integration tests with cargo build.
The binaries of the integration tests appear as [[bin]] binaries in Cargo so that they can be built with rustc in a future script. This means that they are automatically built with cargo build. That is broken since several of them currently lack main functions.

This adds empty main functions to those files and puts the rest of the file content in a submodule marked with #[cfg(test)] to avoid compiler warnings during cargo build. The binaries now do nothing unless compiled with the --test argument.

PiperOrigin-RevId: 497193671