Split the build job in the publish CI workflow into two jobs.
Each job publishes one library. This should reduce the severity of the race condition that the crate googletest_macro has been published on crates.io but is not yet available as a dependency by the time googletest is published. It also allows specifically the job to publish googletest to be restarted (without retrying the already complete publicatio of googletest_macro) in case the race condition does come up.

PiperOrigin-RevId: 499838750