Hide ConjunctionMatcher from generated documentation and explicitly mark it is having an unstable API.
This ensures that it does not show up as part of the API surface in the documentation. It is only public due to a limitation in the Rust language (no impl in return position in trait methods) which will be removed later.

Also adds a short docstring to the trait `AndMatcherExt` itself.

PiperOrigin-RevId: 501825658