Rename the module intersection_matcher to conjunction_matcher.
This matches the convention used in the rest of the library: the module name mirrors the name of the matcher struct within it. The name "intersection_matcher" doesn't correspond directly to anything else in the file.

Also fixes an incorrect reference to the AndMatcherExt::and extension method in the documentation for the all! macro.

PiperOrigin-RevId: 501828070