Display which tuple elements did not match when explaining a mismatch in the tuple! macro.
Previously, the matcher explanation would only show what the elements were expected to be. No information was given on which tuple elements did not match and why.

This implements Matcher::explain_match to enhance the output to show which tuple elements did not match and give explanations for each one.

PiperOrigin-RevId: 503913594