Fold the Describe trait into the Matcher trait.
This simplifies the implementation of matchers, since only one trait must be implemented.

On the downside, testing the describe method directly can be a bit more awkward due to the additional type annotations which are needed. However, this should not be a major problem in practice outside the library itself.

This change breaks API compatibility: existing matchers must be modified. In most cases it suffices to move the implementation of the Describe trait into that of the Matcher trait and remove the Describe implementation and references to it.

PiperOrigin-RevId: 503969016