Fix the use of format string arguments in the fail macro.
The recent change to add #[must_use] to the fail macro put the generated code inside a function. This unfortunately does not work with captured format arguments -- that would require a closure rather than a function.

This change moves the creation of the message out of the function to the function call site, so nothing needs to be captured. It also adjusts the existing test to exercise the aforementioned issue.

PiperOrigin-RevId: 504214102