Use pretty-print to display the actual value in test assertion failure messages.
This makes the output much easier to understand when the actual value consists of a complex data structure with a lot of nesting. Previously, everything was packed into a single line. Pretty-print enforces newlines after every field as well as indentation reflecting the data structure.

This change also fixes several clippy warnings which surfaced in the same files affected by the above modifications.

PiperOrigin-RevId: 505619168