Use pretty-print to display the actual values of individual elements in match explanations.
Previously, the standard debug output (:?) would be used to format individual elements of a collection when building an explanation why a matcher matched or did not match an actual value. Since the standard debug formatter packs the whole output in one line, and individual elements can be complex, this can lead to assertion failure messages which are hard to read.

This changes the output formatter in most cases to use the pretty-print formatter (:#?), which adds newlines and indentation reflecting the data structure.

For now, this change does not touch the display of keys in the has_entry matcher. In most cases, keys should be simple enough that the one-line debug output suffices. Splitting such simple structures into multiple lines in that context may have the opposite of the intended effect.

PiperOrigin-RevId: 505620778