Implement explain_match on remaining matchers which reference inner matchers.
This ensures that a test assertion failure message will contain all additional information pertaining to the actual value to explain the mismatch. Previously, that information would be lost when explain_match was left to its default implementation based on Matcher::describe.

This excludes the has_size matcher. While that references an inner matcher, the only matchers which could be used match directly against numbers, and the only numeric matchers the library has do not implement explain_match themselves. So implementing explain_match on HasSizeMatcher would not bring any advantage and would be quite awkard to test.

PiperOrigin-RevId: 505624541