Do not repeat the actual value inside Option::Some, Result::Ok, or Result::Err in test assertion failure messages.
Previously, the test assertion failure message for the matchers some, ok, and err would contain the actual value. This is redundant since the value was already printed right above. This changes the implementation of explain_match in each of those matchers to just make reference to the explanation of the inner matcher along with a really brief description of the variant found ("has a value", "is a success with value", "is an error with value"). This makes the test assertion failure messages shorter and less jarring.

PiperOrigin-RevId: 505970192