Introduce `StrMatcher`, intended to be a powerful matcher for strings.
This is intended to unify the features of `eq`, `eq_ignoring_ascii_case`, `contains_substring`, `starts_with`, and `ends_with` and to provide a few additional features. Currently, these matchers are completely separate so that one cannot combine their properties. For example, there is no easy way to match a string containing a substring while ignoring ASCII case. The goal of this matcher is to combine all of these into one matcher with a single configuration language to control its behaviour.

This introdces a trait `StrMatcherConfigurator` which allows the developer to configure properties of the string matching, such as ignoring leading and trailing whitespace or ASCII case. The trait is implemented for `EqMatcher<T>` when `T` derefs to `str`, making these configuration options available when one uses `eq(...)`.

This CL implements just one configuration method: `ignoring_leading_whitespace`. That method alone is trivial to implement, allowing this CL to focus on the overall structure of the solution without being bogged down by the actual implementation. While `ignoring_leading_whitespace` by itself would not be terribly useful (one could just call `trim_start()` on the actual and expected values), it may be useful in combination with the configuration method `ignoring_uniform_indentation()` to be introduced later (since calling `trim_start()` would disrupt how that configuration works).

In the future, the goal is to use this matcher for `contains_substring`, `starts_with`, and `ends_with`. The matcher `eq_ignoring_ascii_case` will be removed; one just calls `eq(...).ignoring_ascii_case()` instead. One will also be able to do things like the following:

```
verify_that!(value, eq("
   Some text
   Some text on the next line
").ignoring_uniform_indentation())
```

This allows the actual string value to be indented nicely in the test while still matching it correctly with the actual value.

For users of Kotlin, this is similar to using `trim_indent()` on a string literal and matching by equality.

PiperOrigin-RevId: 510069555