Support string containment in StrMatcher.
This moves the matcher function contains_matcher to the str_matcher module and modifies it to return an StrMatcher. ContainsSubstringMatcher is now obsolete and removed.

Doing this allows one to configure the matcher created by contains_matcher just as one can with eq for string equality: ignoring ASCII case, ignoring uniform indentation, and so forth.

Not all combinations of functionality are supported yet; this includes TODOs at the points which need to be improved.

PiperOrigin-RevId: 510072757