Support ignoring_uniform_indentation with the contains_substring matcher.
This implements an algorithm which performs a line-by-line search for a matching sequence of lines in the actual value. This is complicated by the fact that the initial and final lines are treated as special cases, and a single-line expected value is another special case (although trivial to handle). The resulting implementation keeps a set of iterators through the expected value and terminates when any of those iterators reaches the end.

PiperOrigin-RevId: 510118603