Introduce crate-level documentation.
This will appear in the main page for GoogleTest on docs.rs at https://docs.rs/googletest/*/googletest/index.html. It partly contains a copy of contents from README.md which summarise how to use the library at a high level.

Most importantly, it adds a table of existing matchers, making the matchers much more discoverable. Previously, the documentation of the individual matchers was split between the declarative macros such as all! and field! and the functions. The functions were only accessible by navigating through the matchers module. Now they are all in one place on the front page of the documentation.

PiperOrigin-RevId: 512892231