Add documentation to the declarative macro matchers on what types they accept against what types they match.
This provides some more clarity about the types with which one can use the matchers.

Also adds a couple of tests to ensure that the claims in the documentation are in fact correct.

Also changes some gratuitous uses of the HasSize trait against known Vec's to just call Vec::len.

Fixes #71

PiperOrigin-RevId: 513840639