Fix the matches_pattern! macro.
The last release included a change which redirected matches_pattern! to matches_pattern_internal!. It did not fix the references inside the macros themselves to use $crate::matches_pattern_internal!. This worked fine in contexts where matches_pattern_internal! was already imported, but does not work in general client code where it is not. This change adds the $crate qualifier to each invocation of matches_pattern_internal!.

This also moves the tests of matches_pattern! out of the module and into a new integration test (see https://doc.rust-lang.org/rust-by-example/testing/integration_testing.html). This causes the tests to be compiled from outside the crate itself, precluding this type of bug in the future.

Fixes #86