Add to StrMatcher support for matching a string containing a substring a given number of times.
This is analogous to the `times` method of the `contains` matcher. The tester can now assert, for example, that a given substring appears exactly once and no more. Previously, it was only possible to assert that it appears at least once.

This works in a manor analogous to `times` in `contains`: it adds a field of type `Option<Box<dyn Matcher<usize>>>` to the configuration. If that field is None, then the behaviour is as before. If it is filled, then the matcher is respected and output in the description and match explanation. The use of a trait object avoids having to add more generics to `StrMatcher` just to support this one case, while the use of an `Option` ensures that the description and match explanation aren't affected when the test doesn't specify anything.

PiperOrigin-RevId: 517418605