Introduce a declarative macro to match properties of an object.
This is analogous to the existing macro `field!` which allows applying arbitrary matchers to fields of a struct or enum. This analogously allows matching on the output of method invocations. It is particularly useful in situations where one cannot access the fields directly (due to visibility), or direct access is not desired for other reasons.

This includes two variants of the matcher, one for owned returned values (such as most primitive types) and one for references. The peculiarities of Rust's borrow checking unfortunately require keeping these matchers separate: otherwise the Rust borrow checker does not know how long the returned value will live and cannot prove that it doesn't outlive the data which it references in the outer struct.

Unfortunately, this does not work properly with methods returning string references or slices due to complicated lifetime issues.

PiperOrigin-RevId: 521438328