Support the property matcher macro in matches_pattern.
This allows the keys in a `matches_pattern!` call to include method calls rather than just fields:

```
verify_that!(value, matches_pattern!(MyStruct {
    my_method(): eq(100)
})
```

This is particularly useful when one wants to use `matches_pattern!` in a nested context, but one either cannot access the fields directly due to visibility or one does not want to access the data at that level of abstraction.

This also supports passing arguments to the methods.

Unfortunately, this does not work properly with slices and string references due to complicated issues with lifetimes.

PiperOrigin-RevId: 521441047