Refactor `UnorderedElementsAre` to move as much code as possible into other structs.
In order to allow the macro `unordered_elements_are!` to work with `HashMap` and similar classes, it is necessary to support an `IntoIterator` which iterates over a pair of references. The current implementation iterates over single references. It does not appear to be feasible to add an adapter which makes an iterator over a `HashMap` look like something which can be consumed by `UnorderedElementsAre`.

The solution is to create a new struct to support `HashMap` and friends. In order to minimise duplication of code, as much code as possible should be moved out of `UnorderedElementsAre` into structs which will be in common with both. Fortunately, the vast majority of the code can be moved this way.

PiperOrigin-RevId: 523632451