Support matching HashMap and similar collections in unordered_elements_are!.
This requires adding a new struct UnorderedElementsOfMapAreMatcher which parallels the existing UnorderedElementsAre. Due to limitations in the Rust type system, it does not appear to be possible to unify the implementations further. However, since nearly all of the actual logic is in other structs, the duplication is minimal.

This also analogously renames UnorderedElementsAre to UnorderedElementsAreMatcher, in line with the naming convention used elsewhere in the library.

PiperOrigin-RevId: 523632743