Add support for easily converting anyhow errors into test failures.
This allows more easily returning early from functions returning `anyhow::Error` in conjunction with the `?` operator.

Unfortunately, `anyhow::Error` does not implement `std::error::Error`, so it will not work transparently with the `?` operator. Furthermore, one cannot add a `From<anyhow::Error>` implementation, since theoretically `anyhow::Error` could implement `std::error::Error` in the future, which would then break this crate.

The solution here is to introduce a new trait `IntoTestResult` which is then implemented for `anyhow::Error` and which provides a method `into_test_result()` to do the conversion. The anyhow crate is now an optional dependency enabled by a feature.

This also adds a CI job which tests the library with no default features enabled. This ensures that the library still compiles and works if the downstream dependency does not enable optional dependencies.

PiperOrigin-RevId: 524766870