Rename the atribute macro `google_test` to just `test`.
This changes the canonical way to use the attribute macro from:

```
#[google_test]
fn should_work() -> Result<()> {...}
```

to

```
#[googletest::test]
fn should_work() -> Result<()> {...}
```

This is more in line with other test libraries such as Tokio et al., and aligns with the expectations of libraries like rstest; see https://github.com/la10736/rstest#inject-test-attribute. Thus the test will not be registered twice even if `#[googletest::test]` appears after `#[rstest]`. In particular, this addresses the remainder of https://github.com/google/googletest-rust/issues/77.

This also adjusts the generated output of `#[googletest::test]` to refer explicitly to `::core::prelude::v1::test`, so that it still compiles when the test directly imports `googletest::test`.

Since this is a potentially more serious breaking change, the name `googletest::google_test` is retained as an alias for `googletest::test`. So it is not necessary to port existing downstream uses of the library.

PiperOrigin-RevId: 524774178