Fix support for async tests with `googletest::test`.
Previously, an async test annotated with `googletest::test` would not compile because the test content was put in an ordinary closure. Any uses of `.await` inside that closure would generate a compiler error.

This changes the `googletest::test` macro implementation to detect that the test is async and adjust the generation accordingly. If the test is async, then no closure is generated, but the test content is placed in an async block.

Unfortunately, there does not appear to be a single way to cover both the sync and async cases currently, as explained in the code comments.

A further refinement of this approach could possibly address https://github.com/google/googletest-rust/issues/4.

PiperOrigin-RevId: 524774676