Change the anyhow dependency to support any version in the 1.x series.
Previously, the dependency was locked to 1.0.70 and above, which meant that a downstream library using the anyhow feature would only work if they used that version or a later one in the 1.x series. This is not actually necessary with the use of anyhow in this library. This change relaxes the requirement to accept any version of anyhow with major version 1.

PiperOrigin-RevId: 524816843