Replace the type parameter on Matcher with an associated type.
This should make it possible to associate a lifetime with the actual value in such a way as to allow the property matcher to work with extractors which return data referencing the containing struct.

This unfortunately does not completely work at the moment since some matchers use trait objects and generic associated types are not yet object-safe (see https://blog.rust-lang.org/2022/10/28/gats-stabilization.html#traits-with-gats-are-not-object-safe). This may change in the future.