Remove the use of the `google_test` attribute from unit tests.
This macro expands to code which includes a reference to the `googletest` crate. That reference does not work within the crate itself; the reference would have to be `crate`.

Up until now, this worked because of the alias `use crate as googletest`. That alias will be removed in a follow-up change, which will cause compilation to break without this change.

Since the unit tests don't need the `google_test` attribute and work just fine with `test`, this change just removes that use.

PiperOrigin-RevId: 526005634