Remove the alias crate as googletest from the matchers.
The alias creates problems for doctests, which are compiled in such a context that imports of the form `use googletest::...` become ambiguous.

Making the above change also simplifies the structure of the imports.

Also hid the documentation for the property module, since there are no documented items in it (the only visible item is the macro `property!`, which is documented at the top level).

PiperOrigin-RevId: 526005953