Enable doctests for the library.
This removes the flag which disables doctests in Cargo.toml, so doctests will now run as part of cargo test (and therefore CI).

This change marks all Rust codeblocks in the documentation with `ignore`. Thus this change is de facto a no-op.

Codeblocks which show output -- thus are not code -- have the attribute `text` as per https://doc.rust-lang.org/rustdoc/write-documentation/documentation-tests.html#attributes.

Follow-up changes will modify codeblocks so that they actually work as tests and then enable them.

These changes will allow us to ensure that the code examples in the documentation are accurate.

PiperOrigin-RevId: 526013700