Enable doctests in the matchers::all_matcher module.
Also changes one of the examples in the documentation of the `all!` macro. It turns out that `AndMatcherExt` does not compile when used with matchers on strings, since rustc cannot do the required type inference. Thus the example was changed to use numbers, where the type inference works. It will be necessary to address this issue with `AndMatcherExt`, but that is beyond the scope of this change.

PiperOrigin-RevId: 526556742