Update an obsolete reference in the documentation for `googletest::test`.
This referred to a no-longer-existent type `GoogleTestResult`. This change updates it to refer to the current correct type.

The reference to `googletest::Result` is via a direct URI rather than a crate reference. This is necessary because otherwise googletest_macro would have to depend on googletest to resolve the reference, creating a circular dependency.

PiperOrigin-RevId: 527282097