Move some tests of UnorderedElementsOfMapAreMatcher into the unit tests of the module matchers::unordered_elements_are_matcher.
These tests exercise some of the more internal aspects of the matcher rather than using `verify_that!` with the matcher itself. Because of the type inference issues such use creates, it is more convenient to locate these tests in the unit tests rather than integration tests. This will become particularly apparent in a follow-up change which modifies the `Matcher` trait to use an associated type rather than a type parameter.

PiperOrigin-RevId: 527874814