Consolidate the implementations of the trait `Matcher` for `ContainerEqMatcher`.
Previously, there were several separate implementations to be used against different actual value types. This will no longer be possible in the future when the type parameter of `Matcher` is changed to an associated type.

This change consolidates the various implementations into a single one which already covers all cases. It makes use of the optional type parameter of `PartialEq` to do this.

It also renames the type parameters around `ContainerEqMatcher` to be more descriptive and consistent. The struct itself now has the type of the actual container as a type parameter. This is not strictly required at this stage, but will be necessary in the following step when `Matcher` is modified.

PiperOrigin-RevId: 527875014