Replace the type parameter on Matcher with an associated type.
This has the immediate effect that the extension traits `AndMatcherExt` and `OrMatcherExt` now have no trouble working with strings. Namely, this change forces the matcher struct itself always to reference the actual type against which it matcher, either via parameterisation or by only matching against a concrete type. Because of this, the matcher declaration always carries the necessary type information so that the compiler can perform type inference. This change introduces unit tests which demonstrate this.

This same property has the drawback that matchers must frequently carry a `PhantomData` of the type against which they match. This is annoying, but not fatal.

This is of course an API-incompatible change. Downstream targets which build their own matchers will have to change the type parameter to be an associated type, and may need to make additional adjustments as outlined above. This commit shows several such examples.

This also has the effect that it is no longer possible to define a single matcher class and implement `Matcher` against several actual types with it. Only one matcher in GoogleTest -- `ContainerEqMatcher` -- did this, and a previous change modified it to have only one `Matcher` implementation with no loss of functionality.

Existing tests did not have to be changed, except for tests of some internals of matchers due to direct reference to the `Matcher` type. This suggests that the change will not break downstream _uses_ of matchers.

In the future, this should also make it possible to associate a lifetime with the actual value in such a way as to allow the property matcher to work with extractors which return data referencing the containing struct. Most likely that would require modifying the associated type to have a lifetime parameter. Currently that unfortunately does not work for us since some matchers use trait objects and generic associated types are not yet object-safe (see https://blog.rust-lang.org/2022/10/28/gats-stabilization.html#traits-with-gats-are-not-object-safe). This may change in the future.

PiperOrigin-RevId: 527884316