Improve the example in the documentation for the `and` extension method.
The existing example referenced a struct which is not visible to the reader, so it could not easily be copied out. Making the struct visible would create a great deal more visible clutter.

This PR changes the example to use strings and the `starts_with` and `ends_with` matchers. It also adds some failing examples for more illustration.