Introduce a matcher `eq_deref_of`.
This is similar to `eq` but accepts a reference to the expected value rather than consuming it. This is useful in cases where one wants to check equality, has only a reference to the expected value, and either:
 * the expected value is neither Copy nor Clone, or
 * cloning the expected value is perceived to be too expensive.

This change also refactors the implementation of `explain_match` in `EqMatcher` to allow it to be reused by `EqDerefOfMatcher`.

PiperOrigin-RevId: 529731930