Rename the matcher `size` to `len`.
The name `len` is more idiomatic in Rust for representing the number of elements in a container. The name `size` may be amiguous with the byte size of the data structure.

This also updates the wording in the documentation to refer to "number of elements" rather than the more ambiguous "size" of a container.

This was feedback offered by a participant of the recent Rust Munich Meetup where I talked about GoogleTest Rust.

PiperOrigin-RevId: 531457322