Panic when an `expect_*` assertion runs without the `googletest::test` attribute.
It is too easy to accidentally invoke `expect_that!` whilst forgetting to use the googletest::test attribute macro. This causes the assertion to be silently ignored.

This changes the underlying implementation to panic when there is no test context present and to provide helpful information to the developer in that case. It will also panic if the assertion occurs in a different thread than that in which the test is running. This change should make the use of the library more robust.

The changes will also affect the use of `and_log_failure()` in the same way.

This also fixes the attributes used in some integration and documentation tests, since they were affected by the aforementioned problem.

PiperOrigin-RevId: 531457657