Modify matchers which took references to inner matchers to use `Box` and `Vec`.
This means that the matchers now consistently own rather than borrow their inner matchers. Previously, if one created a matcher with `elements_are!`, `unordered_elements_are!`, or similar in a subroutine, one would run into a compiler error due to the inner matchers being dropped while the outer matcher holds references to them. This limitation is surprising and prevents certain patterns, such as creating a utility function to generate matchers.

This change replaces he references with `Box` and the slice references with `Vec`. Thus the inner matchers are now fully owned by the outer matcher, and there is no problem with dropped temporaries.

This comes at a cost of increased heap allocation, since previously the inner matchers could be allocated entirely on the stack. The cost is fairly small, though, so it's probably still worth it.

PiperOrigin-RevId: 531487537