Fold the extension traits `AndMatcherExt` and `OrMatcherExt` into `Matcher`.
These were previously separate traits so that downstream projects would not depend on `ConjunctionMatcher` or `DisjunctionMatcher` if they did not use that functionality. Since everything is folded into one crate now, that distinction does not exist any more. Folding these into `Matcher` makes them more discoverable and easier to use.

This also adds a re-export of `matcher::Matcher` to the prelude so that downstream code need not explicitly import `Matcher` to use these methods.

Since `and()` and `or()` are no longer supplied by an extension trait, this also removes their entries from the table of matchers.

PiperOrigin-RevId: 531489362