Change the minimum rustc version to 1.59.
This allows GoogleTest Rust to be used with Android, which is using Rust 1.59 for the foreseeable future.

This requires:
 * Removing the use of language features and library APIs which were stabilised after 1.59,
 * Adjusting the dependencies so that they are compatible with Rust 1.59, and
 * Updating the documentation.

The developer experience might not be ideal due prior to Rust 1.66 to some issues with the test harness. In particular, the test failure output may not be as readable as one would like. But the library should be usable.

PiperOrigin-RevId: 531490171