Make `AllMatcher` own its inner matchers.
Previously, `AllMatcher` along with other matchers such as `ElementsAreMatcher` and `UnorderedElementsAreMatcher` would only contain references to their inner matchers. A prior change updated the latter matchers so that they would put these inner matchers in `Box`, making them owned. This change missed `AllMatcher`, the consequence of which is that `AllMatcher` could not be an inner matcher of any of the aforementioned matchers without breaking Rust's borrow checking rules. This also broke any use of `matches_pattern!` as an inner matcher in one of the modified matchers.

This change updates `AllMatcher` to use the same semantics as the other matchers which previously took references to inner matchers, also fixing `matches_pattern!`. It also adds a module of "composition tests" which exercise various matchers in different compositional arrangements to help avoid such problems in the future.