Improve the output of actual value in test assertion failures.
This makes two changes:
 * If the total length of the (non-pretty-printed) debug output of the actual value is less than 60, then the non-pretty-printed output is used. Otherwise, the output is pretty-printed as before. This means that short, structured values like `Some(123)` will no longer be printed with newlines. This should reduce visual clutter and make the actual values easier to read.
 * The match explanation is now separated from the actual value by a newline and is indented. This visually separates the explanation from the actual value, making it easier to distinguish the two.

To support this visual separation, it was also necessary to enhance `Description` to support indenting all but the first line. Otherwise, the `all` matcher would have output multiple entries with the first one indented extra spaces like this:

```
Actual: [1, 2, 3],
    * First explanation
  * Second explanation
  * Third explanation
```

I replaced the boolean field `indented` with a enum representing the indentation mode and cleaned a few things up in the implementation along the way.

This also increases the use of indoc throughout the unit tests, making them easier to read.

PiperOrigin-RevId: 532034798