Support borrowed data inside matchers.
With the previous changes to use `Box` to own matchers in `AllMatcher`, `ElementsAreMatcher`, `UnorderedElementsAreMatcher` and `UnorderedElementsOfMapAreMatcher`, these matchers lost the ability to contain borrowed data with a lifetime other than `'static`. This is because `Box` implicitly assumes that its inner type has lifetime `'static`. In particular, one could not nest `eq_deref_of` inside these matchers. That should not be necessary as long as the `Box` instances are dropped before the references they contain become invalid. Such a limitation is therefore surprising.

This change introduces explicit lifetimes for the content of the `Box` instances. Doing so allows the borrow checker to apply the lifetime of the contained references to prove that they outlive the `Box` instances. This change also introduces integration tests which demonstrate that this works.

PiperOrigin-RevId: 532422248