Restructure `eq_matcher::create_diff` so that it only returns the report on the difference between the strings and not the entire match explanation.
Previously, the method `EqMatcher::explain_match` would defer entirely to `eq_matcher::create_diff`. This made `eq_matcher::create_diff` unusable for `StrMatcher`, where it is intended also to be used. This restructures the function to be reusable by focusing on just the difference between the two input strings.

This also replaces the text "Debug diff" in the match explanation output with "Difference", which should be clearer.

This also adds a parameter to allow the caller to specify the edit distance mode, which will be needed by `StrMatcher`.

PiperOrigin-RevId: 533479136