Support showing a diff of the actual and expected multi-line strings in `StrMatcher`.
This can make the diagnosis of a mismatched string much easier, since it highlights exactly which lines differ. Previously, the actual and expected strings were shown as debug output, making finding the difference extremely tedious.

This is anlogous to the recent work on `EqMatcher`.

Currently this does not support the additional options of `StrMatcher` such as ignoring leading and trailing whitespace. When any of those options are set, the diff output will not appear. There are TODOs in the code to cover this.

PiperOrigin-RevId: 533480053