Move integration tests involving separate binaries into a separate crate.
These are integration tests which test the full behaviour and console output of a test binary using GoogleTest Rust. As such, they (often) rely on extra binaries which must be compiled separately. Keeping these extra binaries in the googletest crate clutters the library itself and increases compilation times to run ordinary tests. This is particularly true when the external binaries depend on additional crates such as rstest and tokio.

This change moves the integration tests outside the googletest crate and into its own crate so that they can be compiled completely independently of googletest. It removes the dependencies (such as rstest and tokio) in the googletest crate which were present only to support the integration tests.